app/backend/requirements.txt
Flask>=2.0
Flask-SQLAlchemy>=3.0
python-dotenv>=1.0
Flask-Cors>=3.0
PyJWT>=2.0
alembic>=1.10
redis>=4.0
passlib[bcrypt]>=1.7
requests>=2.0
rq>=1.1
